Fixing failing tests
