+++ src/live/index.js
@@ -1,0 +1,4 @@
+import LiveServiceManager from './src/manager';
+export {
+  LiveServiceManager
+};
