+++ src/request/index.js
@@ -13,7 +13,6 @@ import Middleware, {
   SerializeMiddleware,
   Storage
 } from './src/middleware';
-import LiveServiceManager from './src/live';
 
 // Export
 export {
@@ -26,7 +25,6 @@ export {
   HttpMiddleware,
   KinveyRequest,
   KinveyResponse,
-  LiveServiceManager,
   MemoryAdapter,
   Middleware,
   NetworkRack,
